Create #newTweetListener div (explanation below)
This div will have a display property of none, so it will be unnoticed
by the user. This div will be given an eventListener that listens for
a "click" - this click will be simulated when generateRandomTweet
executes in the data_generator.js file (further explanation on this
in a future commit). Upon being 'clicked', the event listener attached
to #newTweetListener will fire a callback, which will invoke the
generateTweet function that pushes the new tweet to the DOM.