Create #newTweetListener and fire generateTweet()
When the document is ready (after the page finishes loading), a click
event listener will be attached to the invisible #newTweetListener div,
this event listener will be triggered by the jQuery event that is fired
when a random tweet is generated from data_generator.js. The event
this listner receives will include the newly generated tweet. The
callback on this listener will push the event tweet property to the DOM