Declare tweetSchedulerBool
This bool controls whether the jQuery 'click' event is fired a new
random tweet is generated

Reasoning: the for-loop generates 10 tweets so that when the page loads
tweets are already there. Instead of having these appear one by one
through the simulated jQuery 'click' event, a forEach loop will be
added to twiddler.js to automatically load those tweets. After these
tweets have been loaded, the tweetSchedulerBool will be re-assigned
the value true so that the click event can begin firing and displaying
tweets to the DOM one by one

Additionally: this bool may allow in the future for the display of new
tweets to be paused, but this will potentially be implemented after
the remaining functionality has been added