Refactor userButton / newTweet click listeners
The purpose of this refactorization is to allow showTweets() to know
if its call is as a result of a button being pressed, or if its call
is because a new tweet being generated. If the button is pressed, the
tweet parameter will be undefined in the showTweets() call. On the
other hand, if a tweet is defined in the showTweets() call, then
the function will be able to know that its call is a result of a new
tweet being generated

For userButton:
  1. Declare a variable user and assign the value of the event's
  current target's ID (i.e. the ID of the button that was clicked,
  which contains the username of the button)

For newTweetListener:
  1. Declare a variable tweet and assign it the value of the tweet obj
  2. Declare a variable user and assign it the value of the user
  within the tweet object