Add logic to update ULs in newTweetsListener
Case scenario: the user / tags buttons have loaded on the page, but a
new user tweets, or a new tag is used. This conditional checks to see
if the number of buttons is equal to the sum of keys in the user / tag
objects. If this conditional fails, the lists need to be refreshed.